LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-008-expected.txt


PASS Should use the last fallback position initially
FAIL Should use the third fallback position with enough space left assert_equals: Anchored element should be at the left of anchor expected 135 but got 335
FAIL Should use the second fallback position with enough space above assert_equals: Anchored element should be at the top of anchor expected 135 but got 335
FAIL Should use the first fallback position with enough space left and above assert_equals: Anchored element should be at the left of anchor expected 586 but got 786
